Rename files from define_parameter to define_property
